ar-billboard/requirements.txt
# Core
opencv-python
opencv-contrib-python
numpy
pyyaml
python-dotenv

# (Later steps)
ultralytics
onnxruntime
google-generativeai
